update main function
